Replacing Yammer OAuth with the new OAuth2
