+++ oa-oauth/spec/omniauth/strategies/oauth2/yammer_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
 
 describe OmniAuth::Strategies::Yammer do
-  it_should_behave_like 'an oauth strategy'
+  it_should_behave_like 'an oauth2 strategy'
 end
